FoodAttractivenessSrc/requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
scikit-learn>=1.2.0
Pillow>=9.0.0
PyYAML>=6.0